refactor: changing name Moth to Month
